Refactor apps code into app, manifest, resource, and tool, add additional tests for app_target, some minor function renames, get_app_target refactored to AppTarget::try_from
