+++ crates/apollo-mcp-server/src/apps/mod.rs
@@ -1,0 +1,7 @@
+pub(crate) mod app;
+pub(crate) mod manifest;
+pub(crate) mod resource;
+pub(crate) mod tool;
+
+pub(crate) use app::App;
+pub(crate) use manifest::load_from_path;
