+++ crates/apollo-mcp-server/src/apps/resource.rs
@@ -1,0 +1,412 @@
+use rmcp::ErrorData;
+use rmcp::model::{Meta, Resource, ResourceContents};
+use serde_json::json;
+use url::Url;
+
+use crate::apps::app::{AppResource, AppTarget};
+
+use super::App;
+
+pub(crate) fn get_mime_type(app_target: &AppTarget) -> String {
+    match app_target {
+        AppTarget::AppsSDK => "text/html+skybridge".to_string(),
+        AppTarget::MCPApps => "text/html;profile=mcp-app".to_string(),
+    }
+}
+
+// Attach resource mime type when requested to allow swapping between app targets (Apps SDK, MCP Apps)
+pub(crate) fn attach_resource_mime_type(
+    mut resource: Resource,
+    app_target: &AppTarget,
+) -> Resource {
+    resource.raw.mime_type = Some(get_mime_type(app_target));
+    resource
+}
+
+pub(crate) async fn get_app_resource(
+    apps: &[App],
+    request: rmcp::model::ReadResourceRequestParam,
+    request_uri: Url,
+    app_target: &AppTarget,
+) -> Result<ResourceContents, ErrorData> {
+    let Some(app) = apps.iter().find(|app| app.uri.path() == request_uri.path()) else {
+        return Err(ErrorData::resource_not_found(
+            format!("Resource not found for URI: {}", request.uri),
+            None,
+        ));
+    };
+
+    let text = match &app.resource {
+        AppResource::Local(contents) => contents.clone(),
+        AppResource::Remote(url) => {
+            let response = reqwest::Client::new()
+                .get(url.clone())
+                .send()
+                .await
+                .map_err(|err| {
+                    ErrorData::resource_not_found(
+                        format!("Failed to fetch resource from {}: {err}", url),
+                        None,
+                    )
+                })?;
+
+            if !response.status().is_success() {
+                return Err(ErrorData::resource_not_found(
+                    format!(
+                        "Failed to fetch resource from {}: received status {}",
+                        url,
+                        response.status()
+                    ),
+                    None,
+                ));
+            }
+
+            response.text().await.map_err(|err| {
+                ErrorData::resource_not_found(
+                    format!("Failed to read resource body from {}: {err}", url),
+                    None,
+                )
+            })?
+        }
+    };
+
+    let mut meta: Option<Meta> = None;
+    if let Some(csp) = &app.csp_settings {
+        match app_target {
+            // Note that the difference in which keys are set here and the camelCase vs snake_key is on purpose. These are differences between the two specs.
+            AppTarget::AppsSDK => {
+                meta.get_or_insert_with(Meta::new).insert(
+                    "openai/widgetCSP".into(),
+                    json!({
+                        "connect_domains": csp.connect_domains,
+                        "resource_domains": csp.resource_domains,
+                        "frame_domains": csp.frame_domains,
+                        "redirect_domains": csp.redirect_domains
+                    }),
+                );
+            }
+            AppTarget::MCPApps => {
+                meta.get_or_insert_with(Meta::new).insert(
+                    "csp".into(),
+                    json!({
+                        "connectDomains": csp.connect_domains,
+                        "resourceDomains": csp.resource_domains,
+                        "frameDomains": csp.frame_domains,
+                        "baseUriDomains": csp.base_uri_domains
+                    }),
+                );
+            }
+        }
+    }
+    if let Some(widget_settings) = &app.widget_settings {
+        if let Some(description) = &widget_settings.description
+            && matches!(app_target, AppTarget::AppsSDK)
+        {
+            meta.get_or_insert_with(Meta::new).insert(
+                "openai/widgetDescription".into(),
+                serde_json::to_value(description).unwrap_or_default(),
+            );
+        }
+
+        if let Some(domain) = &widget_settings.domain {
+            meta.get_or_insert_with(Meta::new).insert(
+                match app_target {
+                    AppTarget::AppsSDK => "openai/widgetDomain".into(),
+                    AppTarget::MCPApps => "domain".into(),
+                },
+                serde_json::to_value(domain).unwrap_or_default(),
+            );
+        }
+
+        if let Some(prefers_border) = &widget_settings.prefers_border {
+            meta.get_or_insert_with(Meta::new).insert(
+                match app_target {
+                    AppTarget::AppsSDK => "openai/widgetPrefersBorder".into(),
+                    AppTarget::MCPApps => "prefersBorder".into(),
+                },
+                serde_json::to_value(prefers_border).unwrap_or_default(),
+            );
+        }
+    }
+
+    // In the case of MCP Apps, the meta data is nested under `_meta.ui`
+    if matches!(app_target, AppTarget::MCPApps) {
+        let mut nested = Meta::new();
+        nested.insert("ui".into(), serde_json::to_value(meta).unwrap_or_default());
+        meta = Some(nested);
+    }
+
+    Ok(ResourceContents::TextResourceContents {
+        uri: request.uri,
+        mime_type: Some(get_mime_type(app_target)),
+        text,
+        meta,
+    })
+}
+
+#[cfg(test)]
+mod tests {
+    use rmcp::model::{Extensions, RawResource};
+
+    use crate::apps::manifest::{CSPSettings, WidgetSettings};
+
+    use super::*;
+
+    #[test]
+    fn attach_correct_mime_type_when_open_ai() {
+        let resource = Resource::new(
+            RawResource {
+                name: "TestResource".to_string(),
+                uri: "ui://test".to_string(),
+                mime_type: None,
+                title: None,
+                description: None,
+                icons: None,
+                size: None,
+                meta: None,
+            },
+            None,
+        );
+
+        let mut extensions = Extensions::new();
+        let request = axum::http::Request::builder()
+            .uri("http://localhost?appTarget=openai")
+            .body(())
+            .unwrap();
+        let (parts, _) = request.into_parts();
+        extensions.insert(parts);
+        let app_target = AppTarget::try_from(extensions).unwrap();
+
+        let result = attach_resource_mime_type(resource, &app_target);
+
+        assert_eq!(
+            result.raw.mime_type,
+            Some("text/html+skybridge".to_string())
+        );
+    }
+
+    #[test]
+    fn attach_correct_mime_type_when_mcp_apps() {
+        let resource = Resource::new(
+            RawResource {
+                name: "TestResource".to_string(),
+                uri: "ui://test".to_string(),
+                mime_type: None,
+                title: None,
+                description: None,
+                icons: None,
+                size: None,
+                meta: None,
+            },
+            None,
+        );
+
+        let mut extensions = Extensions::new();
+        let request = axum::http::Request::builder()
+            .uri("http://localhost?appTarget=mcp")
+            .body(())
+            .unwrap();
+        let (parts, _) = request.into_parts();
+        extensions.insert(parts);
+        let app_target = AppTarget::try_from(extensions).unwrap();
+
+        let result = attach_resource_mime_type(resource, &app_target);
+
+        assert_eq!(
+            result.raw.mime_type,
+            Some("text/html;profile=mcp-app".to_string())
+        );
+    }
+
+    #[test]
+    fn attach_correct_mime_type_when_not_provided() {
+        let resource = Resource::new(
+            RawResource {
+                name: "TestResource".to_string(),
+                uri: "ui://test".to_string(),
+                mime_type: None,
+                title: None,
+                description: None,
+                icons: None,
+                size: None,
+                meta: None,
+            },
+            None,
+        );
+
+        let mut extensions = Extensions::new();
+        let request = axum::http::Request::builder()
+            .uri("http://localhost")
+            .body(())
+            .unwrap();
+        let (parts, _) = request.into_parts();
+        extensions.insert(parts);
+        let app_target = AppTarget::try_from(extensions).unwrap();
+
+        let result = attach_resource_mime_type(resource, &app_target);
+
+        assert_eq!(
+            result.raw.mime_type,
+            Some("text/html+skybridge".to_string())
+        );
+    }
+
+    #[test]
+    fn errors_when_invalid_target_provided() {
+        let mut extensions = Extensions::new();
+        let request = axum::http::Request::builder()
+            .uri("http://localhost?appTarget=lol")
+            .body(())
+            .unwrap();
+        let (parts, _) = request.into_parts();
+        extensions.insert(parts);
+        let app_target = AppTarget::try_from(extensions);
+
+        assert!(app_target.is_err());
+        assert_eq!(
+            app_target.err().unwrap().message,
+            "App target lol not recognized. Valid values are 'openai' or 'mcp'."
+        )
+    }
+
+    #[tokio::test]
+    async fn get_app_resource_returns_openai_format_when_target_is_openai() {
+        let app = App {
+            name: "TestApp".to_string(),
+            description: None,
+            resource: AppResource::Local("test content".to_string()),
+            csp_settings: Some(CSPSettings {
+                connect_domains: Some(vec!["connect.example.com".to_string()]),
+                resource_domains: Some(vec!["resource.example.com".to_string()]),
+                frame_domains: Some(vec!["frame.example.com".to_string()]),
+                redirect_domains: Some(vec!["redirect.example.com".to_string()]),
+                base_uri_domains: Some(vec!["base.example.com".to_string()]),
+            }),
+            widget_settings: Some(WidgetSettings {
+                description: Some("Test description".to_string()),
+                domain: Some("example.com".to_string()),
+                prefers_border: Some(true),
+            }),
+            uri: "ui://widget/TestApp#hash123".parse().unwrap(),
+            tools: vec![],
+            prefetch_operations: vec![],
+        };
+
+        let result = get_app_resource(
+            &[app],
+            rmcp::model::ReadResourceRequestParam {
+                uri: "ui://widget/TestApp".to_string(),
+            },
+            "ui://widget/TestApp".parse().unwrap(),
+            &AppTarget::AppsSDK,
+        )
+        .await
+        .unwrap();
+
+        let ResourceContents::TextResourceContents {
+            mime_type, meta, ..
+        } = result
+        else {
+            unreachable!()
+        };
+        assert_eq!(mime_type, Some("text/html+skybridge".to_string()));
+
+        let meta = meta.unwrap();
+        // AppsSDK CSP uses snake_case keys and includes redirect_domains (not base_uri_domains)
+        let csp = meta.get("openai/widgetCSP").unwrap();
+        assert!(csp.get("connect_domains").is_some());
+        assert!(csp.get("resource_domains").is_some());
+        assert!(csp.get("frame_domains").is_some());
+        assert!(csp.get("redirect_domains").is_some());
+        assert!(csp.get("base_uri_domains").is_none());
+        assert!(meta.get("openai/widgetDescription").is_some());
+        assert!(meta.get("openai/widgetDomain").is_some());
+        assert!(meta.get("openai/widgetPrefersBorder").is_some());
+        // AppsSDK should not have ui nesting
+        assert!(meta.get("ui").is_none());
+    }
+
+    #[tokio::test]
+    async fn get_app_resource_returns_mcp_format_when_target_is_mcp() {
+        let app = App {
+            name: "TestApp".to_string(),
+            description: None,
+            resource: AppResource::Local("test content".to_string()),
+            csp_settings: Some(CSPSettings {
+                connect_domains: Some(vec!["connect.example.com".to_string()]),
+                resource_domains: Some(vec!["resource.example.com".to_string()]),
+                frame_domains: Some(vec!["frame.example.com".to_string()]),
+                redirect_domains: Some(vec!["redirect.example.com".to_string()]),
+                base_uri_domains: Some(vec!["base.example.com".to_string()]),
+            }),
+            widget_settings: Some(WidgetSettings {
+                description: Some("Test description".to_string()),
+                domain: Some("example.com".to_string()),
+                prefers_border: Some(true),
+            }),
+            uri: "ui://widget/TestApp#hash123".parse().unwrap(),
+            tools: vec![],
+            prefetch_operations: vec![],
+        };
+
+        let result = get_app_resource(
+            &[app],
+            rmcp::model::ReadResourceRequestParam {
+                uri: "ui://widget/TestApp".to_string(),
+            },
+            "ui://widget/TestApp".parse().unwrap(),
+            &AppTarget::MCPApps,
+        )
+        .await
+        .unwrap();
+
+        let ResourceContents::TextResourceContents {
+            mime_type, meta, ..
+        } = result
+        else {
+            unreachable!()
+        };
+        assert_eq!(mime_type, Some("text/html;profile=mcp-app".to_string()));
+
+        let meta = meta.unwrap();
+        // MCPApps should have ui nesting
+        let ui_meta = meta.get("ui").unwrap();
+        // MCPApps CSP uses camelCase keys and includes baseUriDomains (not redirectDomains)
+        let csp = ui_meta.get("csp").unwrap();
+        assert!(csp.get("connectDomains").is_some());
+        assert!(csp.get("resourceDomains").is_some());
+        assert!(csp.get("frameDomains").is_some());
+        assert!(csp.get("baseUriDomains").is_some());
+        assert!(csp.get("redirectDomains").is_none());
+        assert!(ui_meta.get("domain").is_some());
+        assert!(ui_meta.get("prefersBorder").is_some());
+        // MCPApps should not have description
+        assert!(ui_meta.get("description").is_none());
+    }
+
+    #[tokio::test]
+    async fn get_app_resource_returns_error_for_nonexistent_resource() {
+        let app = App {
+            name: "TestApp".to_string(),
+            description: None,
+            resource: AppResource::Local("test content".to_string()),
+            csp_settings: None,
+            widget_settings: None,
+            uri: "ui://widget/TestApp#hash123".parse().unwrap(),
+            tools: vec![],
+            prefetch_operations: vec![],
+        };
+
+        let result = get_app_resource(
+            &[app],
+            rmcp::model::ReadResourceRequestParam {
+                uri: "ui://widget/NonExistent".to_string(),
+            },
+            "ui://widget/NonExistent".parse().unwrap(),
+            &AppTarget::AppsSDK,
+        )
+        .await;
+
+        assert!(result.is_err());
+    }
+}
